Rename file routes
