+++ src/routes/safe/components/Balances/SendModal/screens/ChooseTxType/index.tsx
@@ -13,7 +13,7 @@ import Paragraph from 'src/components/layout/Paragraph'
 import Row from 'src/components/layout/Row'
 import { safeFeaturesEnabledSelector } from 'src/logic/safe/store/selectors'
 import { useStyles } from 'src/routes/safe/components/Balances/SendModal/screens/ChooseTxType/style'
-import ContractInteractionIcon from 'src/routes/safe/components/Transactions/GatewayTransactions/assets/custom.svg'
+import ContractInteractionIcon from 'src/routes/safe/components/Transactions/TxList/assets/custom.svg'
 
 import Collectible from '../assets/collectibles.svg'
 import Token from '../assets/token.svg'
